ai_agents_course/final_project/ai_agent_step_by_step/04_agent/requirements.txt
# Langflow
# Langflow does not currently publish wheels for Python 3.14+
langflow>=1.0.0

# LLM
litellm>=1.0.0
openai

# RAG
faiss-cpu
rank-bm25
numpy
ollama

# ML/Data Science (for code execution)
pandas
matplotlib
seaborn
scikit-learn
lightgbm
joblib

# Code Execution
subprocess32; python_version < '3.0'

# Utilities
pydantic>=2.0.0
python-dotenv
